♻️ move sonos page builder tests to a dedicated tests file
