+++ tests/jukebox/adapters/inbound/admin/test_ui_controller.py
@@ -250,71 +250,6 @@ def test_ui_controller_registers_fastui_routes_and_page_structure(walk_component
 
 
 @pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_sonos_page_renders_saved_selection_and_discovered_speakers(walk_components):
-    controller = build_controller()
-    route = next(route for route in controller.app.routes if getattr(route, "path", None) == "/api/ui/sonos")
-
-    page = route.endpoint(toast="toast-sonos-success", toast_message="Sonos settings saved.")
-    all_components = list(walk_components(page[0].components))
-
-    assert any(component.type == "Heading" and component.text == "Sonos Speakers" for component in all_components)
-    assert any(component.type == "Heading" and component.text == "Saved selection" for component in all_components)
-    assert any(component.type == "Paragraph" and component.text == "Status: Available" for component in all_components)
-    assert any(
-        component.type == "Paragraph" and component.text == "Coordinator: Living Room [speaker-2]"
-        for component in all_components
-    )
-    assert any(
-        component.type == "Paragraph" and component.text == "Members: Kitchen [speaker-1], Living Room [speaker-2]"
-        for component in all_components
-    )
-    assert any(component.type == "Heading" and component.text == "Discovered speakers" for component in all_components)
-    assert any(component.type == "Paragraph" and component.text == "Kitchen" for component in all_components)
-    assert any(component.type == "Paragraph" and component.text == "Living Room" for component in all_components)
-    assert any(component.type == "Paragraph" and component.text == "Coordinator" for component in all_components)
-    assert any(component.type == "Paragraph" and component.text == "Selected" for component in all_components)
-    assert page[1].event.name == "toast-sonos-success"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_sonos_edit_page_renders_speaker_and_coordinator_selects(walk_components):
-    controller = build_controller()
-    route = next(route for route in controller.app.routes if getattr(route, "path", None) == "/api/ui/sonos/edit")
-
-    page = route.endpoint()[0]
-    form = next(component for component in walk_components(page.components) if component.type == "Form")
-    speakers_field = form.form_fields[0]
-    coordinator_field = form.form_fields[1]
-
-    assert page.components[0].text == "Edit Sonos Selection"
-    assert not any(
-        component.type == "Heading" and component.text == "Saved selection"
-        for component in walk_components(page.components)
-    )
-    assert any(
-        component.type == "Paragraph" and component.text == "Current saved selection"
-        for component in walk_components(page.components)
-    )
-    assert any(
-        component.type == "Paragraph" and component.text == "Coordinator: Living Room [speaker-2]"
-        for component in walk_components(page.components)
-    )
-    assert speakers_field.type == "FormFieldSelect"
-    assert speakers_field.name == "uids"
-    assert speakers_field.multiple is True
-    assert speakers_field.initial == ["speaker-1", "speaker-2"]
-    assert speakers_field.options == [
-        {"value": "speaker-1", "label": "Kitchen (192.168.1.30)"},
-        {"value": "speaker-2", "label": "Living Room (192.168.1.31)"},
-    ]
-    assert coordinator_field.type == "FormFieldSelect"
-    assert coordinator_field.name == "coordinator_uid"
-    assert coordinator_field.initial == "speaker-2"
-    assert form.submit_url == "/api/ui/sonos/edit"
-    assert form.method == "POST"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
 @pytest.mark.anyio
 async def test_update_sonos_selection_saves_and_redirects():
     from jukebox.adapters.inbound.admin.ui_controller import SonosSelectionForm
@@ -375,38 +310,6 @@ async def test_update_sonos_selection_returns_field_error_for_invalid_coordinato
     )
     assert "uids=speaker-1" in response[0].event.url
     assert "coordinator_uid=speaker-2" in response[0].event.url
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_sonos_edit_page_renders_error_banner_and_preserves_submitted_values(walk_components):
-    controller = build_controller()
-    route = next(route for route in controller.app.routes if getattr(route, "path", None) == "/api/ui/sonos/edit")
-
-    page = route.endpoint(
-        error_message="Selected Sonos coordinator must be one of the selected speakers: Living Room [speaker-2]",
-        uids=["speaker-1"],
-        coordinator_uid="speaker-2",
-    )[0]
-    all_components = list(walk_components(page.components))
-    form = next(component for component in all_components if component.type == "Form")
-    speakers_field = form.form_fields[0]
-    coordinator_field = form.form_fields[1]
-
-    assert any(
-        component.type == "Paragraph" and component.text == "Selection not saved" for component in all_components
-    )
-    assert any(
-        component.type == "Paragraph"
-        and component.text == "Selected Sonos coordinator must be one of the selected speakers: Living Room [speaker-2]"
-        for component in all_components
-    )
-    assert speakers_field.initial == ["speaker-1"]
-    assert speakers_field.error is None
-    assert coordinator_field.initial == "speaker-2"
-    assert (
-        coordinator_field.error
-        == "Selected Sonos coordinator must be one of the selected speakers: Living Room [speaker-2]"
-    )
 
 
 @pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
